+++ R/html_print.R
@@ -113,12 +113,22 @@ tab_df <- function(x,
   # get style definition
   style <- tab_df_style(CSS = CSS, ...)
 
-  x <- purrr::map_df(x, function(.i) {
+  if (obj_has_rownames(x)) {
+    rn <- rownames(x)
+  } else {
+    rn <- NULL
+  }
+
+  x <- as.data.frame(lapply(x, function(.i) {
     if (is.numeric(.i) && sjmisc::is_float(.i))
       sprintf("%.*f", digits, .i)
     else
       .i
-  })
+  }))
+
+  if (!is.null(rn)) {
+    rownames(x) <- rn
+  }
 
   # get HTML content
   page.content <- tab_df_content(